pyproject.toml
[tool.poetry]
name = "tmp"
version = "0.1.0"
description = ""
authors = ["Andrei Arkhipov <it.arkhipov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
selene = "2.0.0rc9"
allure-pytest = "^2.13.5"
requests = "^2.32.3"



[tool.poetry.group.dev.dependencies]
ruff = "^0.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
